+++ papis/config.py
@@ -86,12 +86,12 @@ general_settings = {
     "header-format": (
         "<ansired>{doc.html_escape[title]}</ansired>\n"
         " <ansigreen>{doc.html_escape[author]}</ansigreen>\n"
-        "  <blue>({doc.html_escape[year]})</blue> "
+        "  <ansiblue>({doc.html_escape[year]})</ansiblue> "
         "[<ansiyellow>{doc.html_escape[tags]}</ansiyellow>]"
     ),
 
     "info-allow-unicode": True,
-    "ref-format": "{doc[title]:.15}{doc[author]:.6}{doc[year]}",
+    "ref-format": "{doc[title]:.15} {doc[author]:.6} {doc[year]}",
     "multiple-authors-separator": " and ",
     "multiple-authors-format": "{au[family]}, {au[given]}",